% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_column.R
\name{add_time_column}
\alias{add_time_column}
\title{Get Actigraph Counts}
\usage{
add_time_column(time, result, epoch_in_seconds = 1L)
}
\arguments{
\item{time}{a vector of times from the original data to be condensed
using the \code{epoch}}

\item{result}{A \code{data.frame}/matrix with columns \code{X}, \code{Y}, \code{Z}.}

\item{epoch_in_seconds}{epoch to calculate the counts for, in seconds}
}
\value{
A \code{data.frame} of result and the time column
}
\description{
Get Actigraph Counts
}
