% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{floor_unique_seconds}
\alias{floor_unique_seconds}
\title{Make Unique Floored  date/time POSIXct}
\usage{
floor_unique_seconds(x, n = 60L)
}
\arguments{
\item{x}{a vector of date-time objects}

\item{n}{number of seconds}
}
\value{
A \code{POSIXct} vector
}
\description{
Make Unique Floored  date/time POSIXct
}
\examples{
x <- lubridate::ymd_hms("2009-08-03 12:01:29.23")
x = seq(x, x + 2000, by = 1/10)
floor_unique_seconds(x, n = 30)
floor_unique_seconds(x, n = 60)
floor_unique_seconds(x, n = 60*5)
x = as.POSIXlt(x)
floor_unique_seconds(x, n = 60)
floor_unique_seconds(x, n = 60*5)
}
