% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Get Actigraph Counts}
\usage{
get_counts(df, epoch_in_seconds = 1L, sample_rate = NULL)
}
\arguments{
\item{df}{A \code{data.frame} with columns \code{X}, \code{Y}, \code{Z}, and
a time column either \code{time}, \code{HEADER_TIME_STAMP}, or \code{HEADER_TIMESTAMP}}

\item{epoch_in_seconds}{epoch to calculate the counts for, in seconds}

\item{sample_rate}{The sampling rate for the data.  If \code{NULL}, then
it will try to be guessed from \code{df}}
}
\value{
A \code{data.frame} of each axis count and the RMS of them
in the \code{AGCOUNT} column with a time column
}
\description{
Get Actigraph Counts
}
