% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_30_30.R
\docType{data}
\name{raw_30_30}
\alias{raw_30_30}
\title{Raw Accelerometry data}
\format{
A \code{data.frame} with 3 columns, X/Y/Z.
The values are multiplied by 10000 for storage purposes
}
\usage{
raw_30_30
}
\description{
Raw Accelerometry data
}
\keyword{datasets}
