% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_data.R, R/get_counts.R
\name{download_actgraph_test_data}
\alias{download_actgraph_test_data}
\alias{get_counts_py}
\alias{get_counts}
\alias{extract_counts}
\title{Get Actigraph Counts}
\usage{
download_actgraph_test_data()

get_counts_py(
  df,
  epoch_in_seconds = 1L,
  sample_rate = NULL,
  fast = TRUE,
  save_memory = FALSE,
  verbose = TRUE
)

get_counts(
  df,
  epoch_in_seconds = 1L,
  sample_rate = NULL,
  save_memory = FALSE,
  verbose = TRUE
)

extract_counts(raw, epoch_in_seconds = 1L, sample_rate = NULL, verbose = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} with columns \code{X}, \code{Y}, \code{Z}, and
a time column either \code{time}, \code{HEADER_TIME_STAMP}, or \code{HEADER_TIMESTAMP}}

\item{epoch_in_seconds}{epoch to calculate the counts for, in seconds}

\item{sample_rate}{The sampling rate for the data.  If \code{NULL}, then
it will try to be guessed from \code{df}}

\item{fast}{Should the fast implementation be used?  You may want to set
this as \code{FALSE} if your data is \emph{very} big}

\item{save_memory}{Should each column be run separately?}

\item{verbose}{print diagnostic messages}

\item{raw}{a raw matrix of numeric values}
}
\value{
A \code{data.frame} of each axis count and the RMS of them
in the \code{AGCOUNT} column with a time column
}
\description{
Get Actigraph Counts
}
\examples{
x = download_actgraph_test_data()
epoch = 10
sample_rate = 30
search_string = paste0("raw_", epoch, "_", sample_rate)
testfile = x$ActiLifeCounts
testfile = testfile[grepl(search_string, testfile)]
file = x$raw
file = file[grepl(search_string, file)]
if (requireNamespace("readr", quietly = TRUE)) {
   df = readr::read_csv(file, col_names = FALSE)
   colnames(df) = c("Y", "X", "Z")
   df = df[, c("X", "Y", "Z")]
   # title_epoch_frequency
   out_mem = get_counts(df, sample_rate = sample_rate,
   save_memory = TRUE, epoch = epoch, verbose = 2)
   out = get_counts(df, sample_rate = sample_rate, epoch = epoch, verbose = 2)
   out$AGCOUNT = NULL
   check = readr::read_csv(testfile, skip = 10)
   stopifnot(all(check == out))
}
}
